+++ src/state/subscription-status.js
@@ -1,19 +1,49 @@
 import { useState, useEffect } from 'react';
 import { useAPIHandlers } from 'State/api';
+import useStripe from 'State/stripe';
+import { useCurrentUser } from 'State/current-user';
+import { getUserLink } from 'Models/user';
 
 export default function useSubscriptionStatus() {
-  const { getSubscriptionStatus } = useAPIHandlers();
+  const { getSubscriptionStatus, createSubscriptionSession, cancelSubscription } = useAPIHandlers();
+  const stripe = useStripe();
+  const { currentUser } = useCurrentUser();
+  const [{ fetched, isActive, expiresAt }, setSubscriptionStatus] = useState({ fetched: false });
 
-  const [subscriptionStatus, setSubscriptionStatus] = useState({ fetched: false });
-  useEffect(() => {
-    const fetchStatus = async () => {
-      const { data } = await getSubscriptionStatus();
-      setSubscriptionStatus({ ...data, fetched: true });
-    };
-
-    if (!subscriptionStatus.fetched) {
-      fetchStatus();
+  async function subscribe() {
+    try {
+      const { data } = await createSubscriptionSession({
+        successUrl: `${window.location.origin}${getUserLink(currentUser)}`,
+        cancelUrl: `${window.location.origin}/settings`,
+      });
+      const { id: sessionId } = data;
+      stripe.redirectToCheckout({ sessionId });
+    } catch (err) {
+      // TODO decide what kind of error handling we need here
+      console.log(err);
     }
-  });
-  return subscriptionStatus;
+  }
+
+  async function cancel() {
+    try {
+      await cancelSubscription();
+      setSubscriptionStatus({ fetched: true, expiresAt: new Date().toISOString(), isActive: false });
+    } catch (err) {
+      // TODO decide what kind of error handling we need here
+      console.log(err);
+    }
+  }
+
+  useEffect(() => {
+    getSubscriptionStatus().then(({ data }) => {
+      setSubscriptionStatus({ ...data, fetched: true });
+    });
+  }, []);
+  return {
+    fetched,
+    isActive,
+    expiresAt,
+    subscribe,
+    cancel
+  };
 }
